refactor(allocator): split database.py into a db/ package
`database.py` was 1892 lines holding a single `PostgresqlDatabase` god class
with 54 public methods spanning seven unrelated concerns. Every consumer took
a dependency on all of it to use any of it, and the file was too large to hold
in context while editing.

Split it into a `db/` subpackage of focused classes that share one connection
pool, following the pattern `operations_db.py` already established:

    db/__init__.py     deliberately empty (see its docstring)
    db/pool.py    125  PooledCursor, make_pool, validate_pool_sizes, POOL_*
    db/vms.py    1254  VmDatabase — rows, registration/auth, seats, logs, health
    db/schedules.py 203  ScheduleDatabase — scheduled_destructions
    db/metrics.py   220  MetricsDatabase — session-metrics columns
    db/operations.py 212  OperationsDatabase — moved from operations_db.py

No `db/` module imports another except `pool`, so the graph is flat and each
class is readable on its own. `PostgresqlDatabase` is renamed `VmDatabase` to
match its three role-named siblings.

Also removes the psycopg2 sys.modules mocking apparatus
---------------------------------------------------------
`operations_db.py` carried a 15-line docstring explaining why it had to
lazy-import `_PooledCursor`: the test suite mocked psycopg2 by *captured
binding*, patching `sys.modules` around an import so the module under test
permanently held the mock. That made import order load-bearing, and moving pool
construction between modules broke 24 tests.

Replaced with dependency injection — `VmDatabase` accepts an optional `pool=`
and tracks ownership, so `__del__` never closes a pool it was handed. Tests now
inject a `MagicMock` pool. No `sys.modules` psycopg2 mocking remains anywhere in
the allocator suite, and the lazy-import workaround is gone.

Two conventions this creates, both documented in the code:
  - `db/__init__.py` stays empty; re-exporting would make importing any
    submodule execute `db/vms.py`'s top-level `import psycopg2`.
  - A class that *constructs* a pool must do so in its own module, so its own
    psycopg2 binding is used. `make_pool` is for bare-pool callers.

Dead code removed
-----------------
Six methods reachable only from their own tests, plus the 9 tests covering
them: `insert_vm`, `get_vm_by_machine_identity`, `get_gpu_health`,
`save_logs_by_hostname`, `load_database`, `seal_session_metrics`.
(`bulk_seal_session_metrics` is live and retained.)

Other fixes surfaced along the way
----------------------------------
  - `execute_scheduled_destruction_job` built a whole `PostgresqlDatabase` just
    to reach `.pool`, and ended in a `finally` block testing `.cursor`/`.conn`,
    attributes that had not existed since an earlier pool refactor. It now
    shares one pool across all four handles, with a reachable `closeall()`.
  - The two duplicated 13-element scheduled_destructions column lists are
    collapsed into one `_SCHEDULE_COLUMNS` tuple.
  - `operations_db.py`'s docstring claimed `scheduled_destructions` reaches into
    VM rows; it does not, at the schema level.
  - Several test fixtures shared one MagicMock across two collaborators, so they
    no longer pinned which object a call landed on. Split into distinct doubles.

No SQL, schema, or runtime behaviour changes. 751 passed, 19 skipped;
`ruff check src tests` clean.
